Added lmtsh man page
Added an lmtsh man page consisting of the usage information that is
printed by "lmtsh -h".  The man page is in section 8 because it can only
be run by root, and, as a result, it fits under the category of System
adminstration commands. Added the new file to the configure.ac and
Makefile.am files so that it will be included in the build.  Also added
the generated man page to scripts/.gitignore.

Closes #15