Fixed head and tail operations on OSS_DATA table
Fixed head and tail operations on the OSS_DATA table.  Added a condition
to check whether the table is OSS_DATA when the operation is head or
tail.  If so, it sets up the query to get the first or last n rows of
the table.  Tested that head and tail operations work on OSS_DATA table.
Also tested that head and tail operations work on all tables which they
worked on before the change.

Closes #18