Sort the "CR" column correctly in ltop
The ltop "CR" column should be sorted in a descending numerical order.
However, often that is not the case. From the code, I note that the CR
column uses sample_rate() to display the number but sample_val_cmp()
to sort the numbers. This commit changes the latter to be sample_rate_cmp().
Based off of randomly generated data, the sorting now works as expected.