Fix compilation against Lua 5.2.
Lua 5.1 replaced the function lua_open with a preprocessor macro which
called luaL_newstate.  Lua 5.2 then removed lua_open entirely.  Thus,
LMT won't build against Lua 5.2.  This change replaces the call to
lua_open() with a call to luaL_newstate(), and since that function only
exists in Lua >= 5.1, a minimum version number of 5.1 was added to the
spec file BuildRequires.