Fix _find_all_fs to find all filesystems
Modified _match_mdtstat() and _match_oststat() to compare both the
numeric portion of the target name and the filesystem name.  With this
change, mdt records with the same numeric identifier, but for different
filesystems, get separate entries in the mdt_data list.  The same holds
for the ost_data list.  _find_all_fs() builds its list fsl from the fsname
field of each mdt_data list element, and the ost counts by counting
entries in the ost_data list with matching fsname values.