Report recovery status for MDT
Update the protocol to include MDT recovery status.

Define a version 2 of the lmt_mdt message, which includes the
recovery status field.  Obtain the recovery status for MDTs from
the procfile recovery_status, as is already done for OSTs.

Move code to parse the procfile out of ost.c to enable it to be
used by mdt.c as well.

Define a parser for the lmt_mdt version 2 message, and modify ltop
to use the appropriate version.  Similarly update the cerebrod
monitor, lmt_mysql.

When an MDT reports it is in RECOVERY, display a message in ltop
at the top of the window to the right of the filesystem name.  If
more than one MDT is in recovery, the full status of the first one
encountered will be displayed.  This allows the user to determine
whether the count of re-connected clients is increasing or not.

Update the tests to reflect the protocol change.