ltop: fix buffer overflow in recovery status
A buffer used to display recovery status from one of the MDTs
could be overflowed when the recovery status was copied from
the cerebro message.

The buffer was also too small to contain the entire recovery
status, so that the status visible to the user was sometimes
truncated.

This patch fixes both issues.