Create separate sub-window listing each MDT
Create a separate sub-window which lists each MDT, as already existed
for OSTs.  Like the existing OST sub-window, the MDT sub-window supports
sorting, collapsing MDTs to MDSs, and scrolling when there are more rows
to display than rows available on the terminal.

For scrolling and selecting the sort column, the user switches between
the MDT and OST sub-windows via 'z' or 'Z'.  This, and the new keys
introduced for selecting an MDT sort column, are listed in the help
display.

If room is available in the terminal window, the summary as well as rows
for each MDT and OST are displayed.

Otherwise, the MDT sub-window and OST sub-window sizes are adjusted to
fit in the available space.  The MDT and OST sections each occupy a
number of rows, respectively, proportional to the MDT and OST counts.

Updated manpage to reflect new sub-window and keys.