In ltop ensure MDT and OST windows have 3 rows
If there is room for ltop to display individual MDTs and OSTs, always
give each of those windows at least 3 rows - 1 for the header and 2 for
the targets or servers.

Fixes a bug on servers with more MDTs than OSTs, where the OST window
might be given only 1 row, which would display the header, and have no
room to show targets.