libproc: look in /sys and /proc as necessary
Rather than always looking under /proc for a given subdirectory,
look under /sys and then if that is not found, look under /proc.
Use whichever path is found first.

This was done to allow lmt to find Lustre 2.12 metrics, some of
which were moved to /sys, but with the path under /sys the same
as the old path under /proc.

This adds some stat() calls, increasing the I/O cost of the the
metrics; but since they are querying /sys, they are entirely
local and should be inexpensive.

Closes #35

Reviewed-by: Christopher J. Morrone <morrone2@llnl.gov>