Recovery status reported for any status other than COMPLETE
According lprocfs_recovery_status_seq_show() in Lustre 2.13,
valid "status" values in recovery_status are:
COMPLETE             The target is active and handling requests
WAITING              The target is active but waiting for another MDT
WAITING_FOR_CLIENTS  The target is active but no clients have connected
RECOVERY             The target is active and recovering after failover
INACTIVE             The target is inactive

For individual targets, for all states other than COMPLETE, display the
recov_status field instead of metric values.   This makes it easier for
the admin to see unhealthy targets.

At the top of the window, report the lowest-numbered MDT which is not
COMPLETE or INACTIVE.  If an MDT is INACTIVE, it was set that way by
an admin and she likely already knows - but other states may not be
expected and should be brought to her attention.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>