Revert "Add /sys in Lustre version checking (#31)"
This reverts commit 8a2a19c70f565962d9d19412e9978de0bed41fc5.

8a2a19c causes lmt to always look in the real /proc to find the lustre
version, which then influences where metrics are found and how they
are parsed.  This defeats the compatability tests performed via "make
check" which explicitly provide lmtmetric with a path to /proc trees
representing past Lustre versions (see test/lustre_versions).

In addition, a later commit,
* 42db6b8 libproc: look in /sys and /proc as necessary

handles the one of the issues 8a2a19c was written for, but also handles
it for other metrics moved to /sys, rather than 8a2a19c which handles
only the Lustre version.

The other issue 8a2a19c was written for is that the contents of
/sys/fs/lustre/version used to be in a key-value format with more than
just a bare version string, but is just a bare version string as of
Lustre 2.10.  That issue is handled by a separate commit.

Conflicts:
	libproc/lustre.c

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>