handle lustre version file containing bare version
Up to Lustre 2.8, the running lustre version could be found in
/proc/fs/lustre/version, formatted like this:

	lustre: 2.0.56
	kernel: patchless_client
	build:  2.0.56.0-gf83e7db-CHANGED-2.6.18-194.17.1.el5_lustre.gf83e7db

As of Lustre 2.10, the version is found in /sys/fs/lustre/version and
contains a bare version string, like this:

	2.10.8_5.chaos

This patch alters _read_lustre_version_string() so that it parses both
types of content correctly.

Add an error check in _find_mdt_dir() to correctly detect when
_read_lustre_version_string() returns an error rather than a packed
lustre version.

Fixes #31

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Reviewed-by: Tony Hutter <hutter2@llnl.gov>