Remove err_exit() calls from libproc
Earlier commit
* f984897 handle lustre version file containing bare version

introduced a bug by adding err_exit() calls to
libproc/{mdt,ost,osc,router}.c call paths.

This caused the cerebro metric modules _get_metric_value() calls to
return -1 (error) when the metrics were run on a system without the
relevant lustre subsystem running.  An example is a Lustre server after boot but
before starting MDTs or OSTs.  This is incorrect, as the subsystem not
running is a valid state - the module can return no metric, at least for
string type metrics, and indicate success.

In addition, returning -1 triggers a bug in cerebrod, and cerebrod exits.

Remove the calls, and update the test expected output to reflect the errors
reported now that the functions are running to completion.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>