Error handling fixes in lustre version code (#52)
Remove calls to msg_exit() in code paths for reading and parsing the
lustre version.  Errors in these code paths may occur because the
lustre subsystem target is simply not started, which itself is not an
error.  In this cases, cerebro metric modules must return 0 for
success.

In some of these code paths, some functions returned an rc < -1.
Alter these functions to return -1 instead, for consistency with the
rest of the code.

Fixes bugs introduced by commits:
* f984897 handle lustre version file containing bare version
* 91c9c9d Revert "Add /sys in Lustre version checking (#31)"

Closes #52

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Reviewed-by: Tony Hutter <hutter2@llnl.gov>