add mdt metric version 3 for DoM
The read and write bandwidth for MDTs are
added to support filesystems using the Dom
(data on MDT) feature.

The values for these 2 new parameters are appended to
the end of the version 2 mdt metric string. This results in
an incompatibility due to the length of the metric string
being significant, so a new version is created, version 3.

ltop is modified so that Rename and Close are removed from
the per-MDT stats, and are replaced by rMB/s and wMB/s.

The database portion of lmt has been updated enought to compile,
but has not been tested.

The test code has been updated to include a test for mdt_v3.