add lmtinit flag to update opnames
Add option to update the mdt operation names
in exiting databases. This is necessary due to
the new op names introduced by mdt_v3.

When cerebrod uses lmt_mysql to insert data into
table MDS_OPS_DATA, the data will be rejected unless
the OPERATION_ID is associated with an OPERATION_NAME
in table OPERATION_INFO.

Previously, the OPERATION_NAMEs and OPERATION_IDs were
set only at database creation time.

This change allows users to add all of the operation names
from the 2 most recent versions of the mdt metric
to OPERATION_INFO. This allows data for the new opnames
to be stored in MDS_OPS_DATA.

This is not a database schema change.