fix typo in name of lmt_db_insert_mdt_v1_v2_v3
The function mt_db_insert_mdt_v1_v2 was renamed
to lmt_db_insert_mdt_v1_v2_v3, but the final 'v'
was left out of the name in the function definition.
It was spelled correctly in places where the function
was called. This resulted in a compilation error
for part of the database code.