remove mdt_v1 and mds_v2 compatibility
With the addition of mdt_v3, the two newest
versions of the mdt metric are v2 and v3.
mdt_v1 and mds_v2 (which precedes mdt_v1) can
be removed. This change makes it so that ltop
and the database daemon will reject mdt_v1
and mds_v2 when processing data from cerebro
and from ltop recordings.