check that lustre exists for lmt_mdt
Currently, lmt_mdt checks the lustre version to
determine if it should look in an mds or mdt directory.
However, to do so it checks
/{sys,proc}/fs/lustre/version which will not exist
if lustre isn't loaded. This will cause it
to complain in the logs every cerebro heartbeat.

Have lmt_mdt check that /{sys,proc}/fs/lustre/ exists
before attempting to get the version. If no lustre
directories are found, assume that lustre is not
installed and exit silently.