change ltop window sizing algorithm
ltop now has a new window sizing algorithm
that prioritizes either OSTs or MDTs.
All MDTs or OSTs are shown before allocating space
for the other server category.
The original algorithm, which allocated window space
proportional to the number of total MDTs and OSTs,
is still available, but not used by default.