lmtinit: fix argument handling
Correct the optstring getopt uses to know which options take
arguments.  The x option for dumping the config does not take
an argument, so it should not have a : after it.

When argument parsing fails, emit an error message indicating the
nature of the problem in addition to the usage message, so the
user knows what to look for.

Fixes #67

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>