ltop: make LOCKS, LGR, LCR columns in the OST window wider
Increase the width of these fields, they aren't wide enough to keep
the table aligned on garter with mdtest on 100 mutt nodes against one OST:

>OST S        OSS   Exp   CR rMB/s wMB/s  IOPS   LOCKS  LGR  LCR %cpu %mem %spc
0000      garter5   122    0     0     0     0       0 10267 10268    5   77   18
0001      garter6   122    0     0     0     0       0  202  203    6   63   20
0002      garter7   122    0     0     0     0       0 11995 11996    8   91   17
0003      garter8   122    0     0   365  6790 23018655 9999 6578   12   82   22

The MDT window was already widened when we added rMB/s and wMB/s with the
advent of data-on-mdt.  The OST window is still narrower than the MDT window.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>